refactor: Architecture SOLID pour TrackedEmailsTable
Restructuration complète du composant TrackedEmailsTable selon les principes SOLID :

**Single Responsibility Principle**
- Extraction des utilitaires de formatage (email-formatting.ts)
- Création de hooks spécialisés (data, actions, filters, table)
- Séparation des composants UI (filters, bulk-actions, pagination)
- Extraction des définitions de colonnes

**Open/Closed Principle**
- Composants extensibles via props et composition
- Configuration via objets d'options

**Liskov Substitution Principle**
- Interfaces cohérentes pour tous les hooks
- Types génériques réutilisables

**Interface Segregation Principle**
- Hooks focalisés sur une seule responsabilité
- Pas de dépendances inutiles

**Dependency Inversion Principle**
- Dépendance sur abstractions (hooks) plutôt que implémentations concrètes
- Injection de dépendances via props

Structure créée :
- lib/utils/email-formatting.ts (formatage de dates, destinataires)
- components/tracked-emails/hooks/ (useTrackedEmailsData, useEmailActions, useEmailFilters, useTrackedEmailsTable)
- components/tracked-emails/columns/ (TrackedEmailsColumns.tsx)
- components/tracked-emails/filters/ (TrackedEmailsFilters.tsx)
- components/tracked-emails/bulk-actions/ (TrackedEmailsBulkActions.tsx)
- components/tracked-emails/pagination/ (TrackedEmailsPagination.tsx)
- TrackedEmailsTable.tsx réduit à un composant d'orchestration (~190 lignes vs ~984 lignes)

Bénéfices :
- Code plus maintenable et testable
- Composants réutilisables
- Responsabilités clairement définies
- Facilite l'ajout de nouvelles fonctionnalités
