+++ components/tracked-emails/pagination/TrackedEmailsPagination.tsx
@@ -1,0 +1,160 @@
+/**
+ * TrackedEmailsPagination Component
+ *
+ * Provides pagination controls for the tracked emails table.
+ * Follows Single Responsibility Principle - handles only pagination UI.
+ */
+
+"use client";
+
+import { useId } from "react";
+import { Table } from "@tanstack/react-table";
+import {
+  ChevronFirstIcon,
+  ChevronLastIcon,
+  ChevronLeftIcon,
+  ChevronRightIcon,
+} from "lucide-react";
+import { Button } from "@/components/ui/button";
+import { Label } from "@/components/ui/label";
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select";
+import {
+  Pagination,
+  PaginationContent,
+  PaginationItem,
+} from "@/components/ui/pagination";
+import type { TrackedEmailWithDetails } from "@/lib/types";
+
+export interface TrackedEmailsPaginationProps {
+  table: Table<TrackedEmailWithDetails>;
+}
+
+/**
+ * Pagination component for tracked emails table
+ * @param table - TanStack Table instance
+ */
+export function TrackedEmailsPagination({
+  table,
+}: TrackedEmailsPaginationProps) {
+  const id = useId();
+  const pageState = table.getState().pagination;
+
+  const startRow = pageState.pageIndex * pageState.pageSize + 1;
+  const endRow = Math.min(
+    (pageState.pageIndex + 1) * pageState.pageSize,
+    table.getRowCount()
+  );
+
+  return (
+    <div className="flex items-center justify-between gap-8">
+      {/* Results per page selector */}
+      <div className="flex items-center gap-3">
+        <Label htmlFor={id} className="max-sm:sr-only">
+          Lignes par page
+        </Label>
+        <Select
+          value={pageState.pageSize.toString()}
+          onValueChange={value => {
+            table.setPageSize(Number(value));
+          }}
+        >
+          <SelectTrigger id={id} className="w-fit whitespace-nowrap">
+            <SelectValue placeholder="Nombre de résultats" />
+          </SelectTrigger>
+          <SelectContent className="[&_*[role=option]]:ps-2 [&_*[role=option]]:pe-8 [&_*[role=option]>span]:start-auto [&_*[role=option]>span]:end-2">
+            {[5, 10, 25, 50].map(size => (
+              <SelectItem key={size} value={size.toString()}>
+                {size}
+              </SelectItem>
+            ))}
+          </SelectContent>
+        </Select>
+      </div>
+
+      {/* Page number information */}
+      <div className="text-muted-foreground flex grow justify-end text-sm whitespace-nowrap">
+        <p
+          className="text-muted-foreground text-sm whitespace-nowrap"
+          aria-live="polite"
+        >
+          <span className="text-foreground">
+            {startRow}-{endRow}
+          </span>{" "}
+          sur{" "}
+          <span className="text-foreground">
+            {table.getRowCount().toString()}
+          </span>
+        </p>
+      </div>
+
+      {/* Pagination navigation buttons */}
+      <div>
+        <Pagination>
+          <PaginationContent>
+            {/* First page button */}
+            <PaginationItem>
+              <Button
+                size="icon"
+                variant="outline"
+                className="disabled:pointer-events-none disabled:opacity-50"
+                onClick={() => table.firstPage()}
+                disabled={!table.getCanPreviousPage()}
+                aria-label="Première page"
+              >
+                <ChevronFirstIcon size={16} aria-hidden="true" />
+              </Button>
+            </PaginationItem>
+
+            {/* Previous page button */}
+            <PaginationItem>
+              <Button
+                size="icon"
+                variant="outline"
+                className="disabled:pointer-events-none disabled:opacity-50"
+                onClick={() => table.previousPage()}
+                disabled={!table.getCanPreviousPage()}
+                aria-label="Page précédente"
+              >
+                <ChevronLeftIcon size={16} aria-hidden="true" />
+              </Button>
+            </PaginationItem>
+
+            {/* Next page button */}
+            <PaginationItem>
+              <Button
+                size="icon"
+                variant="outline"
+                className="disabled:pointer-events-none disabled:opacity-50"
+                onClick={() => table.nextPage()}
+                disabled={!table.getCanNextPage()}
+                aria-label="Page suivante"
+              >
+                <ChevronRightIcon size={16} aria-hidden="true" />
+              </Button>
+            </PaginationItem>
+
+            {/* Last page button */}
+            <PaginationItem>
+              <Button
+                size="icon"
+                variant="outline"
+                className="disabled:pointer-events-none disabled:opacity-50"
+                onClick={() => table.lastPage()}
+                disabled={!table.getCanNextPage()}
+                aria-label="Dernière page"
+              >
+                <ChevronLastIcon size={16} aria-hidden="true" />
+              </Button>
+            </PaginationItem>
+          </PaginationContent>
+        </Pagination>
+      </div>
+    </div>
+  );
+}
